pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "colorlog>=6.9.0",
    "fastapi>=0.116.1",
    "fastapi-derive-responses>=0.1.4",
    "fastapi-swagger>=0.3.29",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.11.7",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "ruff>=0.12.11",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 120
lint.ignore = [
    # wildcard imports are fine
    "F403",
    "F405",
    "PLR"
]
lint.extend-select = ["I", "UP", "PL"]
exclude = [
    ".git",
    ".venv",
    "venv",
    "alembic",
]
